font/segoe/requirements.txt
fonttools>=4.40.0
Pillow>=9.0.0 